+++ packages/granularity/src/components/GrSidebar/config.ts
@@ -1,6 +1,9 @@
 import { defineGranularComponent } from '@feugene/unocss-preset-granular/contract'
 
+import { grSidebarSafelist } from './safelist'
+
 export const grSidebarConfig = defineGranularComponent(import.meta.url, {
   name: 'GrSidebar',
   dependencies: ['GrButton', 'GrIcon'],
+  safelist: grSidebarSafelist,
 })